# apidoc.yaml
openapi: '3.0.2'
info:
  title: Spring Boot API
  version: '1.1'
  description: API Description
  contact:
    name: API Support
    email: myemail@email.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: https://localhost:8080/api/v1
    description: Localhost server
  - url: https://api.example.com/api/v1
    description: Production server
  - url: https://api.staging.example.com/api/v1
    description: Staging server
tags:
  - name: books
    description: Books from library
  - name: users
    description: Users from library
paths:
  /library/users:
    get:
      summary: Get users from library
      description: Get users from library with optional filters by name and email
      operationId: getUsers
      tags:
        - "users"
      parameters: 
        - name: name
          in: query
          description: User name
          required: false
          schema:
            type: string
        - name: email
          in: query
          description: User email
          required: false
          schema:
            type: string
        - name: age
          in: query
          description: User age
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
    post:
      summary: Create user
      description: Create user with name, email and age
      operationId: createUser
      tags:
        - "users"
      requestBody:
        description: User object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created user successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
  /library/users/{id}:
    put:
      summary: Update user
      description: Update user with name, email and age
      operationId: updateUser
      tags:
        - "users"
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: User object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
    delete:
      description: Delete user
      operationId: deleteUser
      tags:
        - "users"
      parameters:
        - name: id
          in: path
          description: User id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
  /library/books:
    get:
      summary: Get books
      description: Get books with optional filters by name and author
      operationId: getBooks
      tags:
        - "books"
      parameters: 
        - name: name
          in: query
          description: Book name
          required: false
          schema:
            type: string
        - name: author
          in: query
          description: Book author
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Error400'
    post:
      summary: Create book
      description: Create book with name, author, ISBN and description
      operationId: createBook
      tags:
        - "books"
      requestBody:
        description: Book object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '201':
          description: Created book successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
  /library/books/{id}:
    put:
      description: Update book
      operationId: updateBook
      tags:
        - "books"
      parameters:
        - name: id
          in: path
          description: Book id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Book object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
    delete:
      description: Delete book
      operationId: deleteBook
      tags:
        - "books"
      parameters:
        - name: id
          in: path
          description: Book id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'

# objects
components:
  schemas:
    BeerResponse:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: Punk IPA 2007
        tagline:
          type: string
          example: Post Modern Classic. Spiky. Tropical. Hoppy.
        first_brewed:
          type: string
          example: 04/2007
        description:
          type: string
          example: Our flagship beer that kick started the craft beer revolution. This is James and Martin's original take on an American IPA, subverted with punchy New Zealand hops. Intense pine and kiwi fruit coupled with a big hop hit.
        image_url:
          type: string
          example: https://images.punkapi.com/v2/192.png
        abv:
          type: number
          example: 6.0
        ibu:
          type: number
          example: 60.0
        ebc:
          type: number
          example: 20.0
    UserRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@email.com
        age:
          type: integer
          format: int32
          example: 30
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@email.com
        age:
          type: integer
          format: int32
          example: 30
    BookRequest:
      type: object
      properties:
        name:
          type: string
          example: The Hobbit
        author:
          type: string
          example: J.R.R. Tolkien
        isbn:
          type: string
          example: 978-3-16-148410-0
        description:
          type: string
          example: The Hobbit is a tale of high adventure, undertaken by a company of dwarves in search of dragon-guarded gold.
    BookResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        author:
          type: string
        isbn:
          type: string
        description:
          type: string
    Error400:
      type: object
      properties:
        message:
          type: string
          example: Bad Request
        code:
          type: integer
          format: int32
          example: 400
    Error404:
      type: object
      properties:
        message:
          type: string
          example: Not Found
        code:
          type: integer
          format: int32
          example: 404
        
    